chore: removed reference infos app
