+++ base_static/global/base.css
@@ -7,7 +7,7 @@
 
     body{
         background-color: cyan;
-        font-family: Arial, Helvetica, sans-serif;
+        font-family: Roboto, Arial, Helvetica, sans-serif;
     }
 
     header{
