+++ base_templates/global/base.html
@@ -11,7 +11,7 @@
     <header>
         <h1><a href="/" class="nav-title">Pense Nisso</a></h1>
         <ul>
-            <li><a href="{% url 'infos:explorer'%}">Empresas</a></li>
+            <li><a href="{% url 'company:explorer'%}">Empresas</a></li>
         </ul>
     </header>
     <hr>
