+++ company/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from company.models import Company
+
+# Register your models here.
+admin.site.register(Company)
