+++ company/models.py
@@ -1,0 +1,15 @@
+from django.db import models
+
+# Create your models here.
+
+class Company(models.Model):
+    name = models.CharField(max_length=100)
+    description = models.TextField()
+    logo = models.ImageField(upload_to='company/logo/%y/%m/%d/')
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    def __str__(self):
+        return self.name
+    class Meta:
+        verbose_name_plural = "Companies"
+
