+++ company/templates/companies/company.html
@@ -9,9 +9,9 @@
     </div>
     <div>
         <ul>
-            <li><a href="{% url 'infos:reports' company.id %}">Denúncias</a></li>
-            <li><a href="{% url 'infos:lawsuits' company.id %}">Processos</a></li>
-            <li><a href="{% url 'infos:news' company.id %}">Notícias</a></li>
+            <li><a href="{% url 'company:reports' company.id %}">Denúncias</a></li>
+            <li><a href="{% url 'company:lawsuits' company.id %}">Processos</a></li>
+            <li><a href="{% url 'company:news' company.id %}">Notícias</a></li>
         </ul>
     </div>
 {% endblock %}
