--- infos/models.py
@@ -1,49 +1,0 @@
-from django.db import models
-
-# Create your models here.
-
-class ReportCategory(models.Model):
-    name = models.CharField(max_length=100)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    def __str__(self):
-        return self.name
-
-class News(models.Model):
-    title = models.CharField(max_length=100)
-    content = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    company = models.ForeignKey('Company', on_delete=models.CASCADE)
-    def __str__(self):
-        return self.title
-
-class Report(models.Model):
-    title = models.CharField(max_length=100)
-    content = models.TextField()
-    category = models.ForeignKey('ReportCategory', on_delete=models.SET_NULL, null=True)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    links = models.TextField()
-    company = models.ForeignKey('Company', on_delete=models.CASCADE)
-    def __str__(self):
-        return self.title
-
-class Lawsuit(models.Model):
-    title = models.CharField(max_length=100)
-    content = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    links = models.TextField()
-    company = models.ForeignKey('Company', on_delete=models.CASCADE)
-    def __str__(self):
-        return self.title
-
-class Company(models.Model):
-    name = models.CharField(max_length=100)
-    description = models.TextField()
-    logo = models.ImageField(upload_to='images/')
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    def __str__(self):
-        return self.name
